Split Plot3DActions into actions subpackage as in plot
